Updated image URLs to have a relative path
In order for the web application to be able to locate the images, they need to have a path URL that is relative to where the index.html file is located.  File paths that start with a slash are absolute (the starting slash means the root folder of the file system, so C: if you are on windows, or the root folder of the hard drive on Mac or Linux).  File paths that start with a name (no slash) are relative to the current directory